+++ tests/Fixtures/TestBundle/GraphQl/Type/Definition/UploadType.php
@@ -30,6 +30,7 @@ final class UploadType extends ScalarType implements TypeInterface
      * @var string
      */
     public $name = 'Upload';
+
     /**
      * @var string
      */