+++ tests/GraphQl/Action/EntrypointActionTest.php
@@ -26,6 +26,7 @@ use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\PropertyAccess\PropertyAccess;
 use Symfony\Component\Routing\RouterInterface;
 use Twig\Environment as TwigEnvironment;
 
@@ -82,26 +83,31 @@ class EntrypointActionTest extends TestCase
         $this->assertEqualsWithoutDateHeader(new JsonResponse(['GraphQL']), $mockedEntrypoint($request));
     }
 
-    public function testFormMultipartAction()
+    /**
+     * @dataProvider multipartFormProvider
+     */
+    public function testMultipartFormAction(?string $operations, ?string $map, array $files, array $variables, Response $response)
     {
-        $file = new UploadedFile(
-            __DIR__.'/Fixtures/test.gif',
-            'original.gif',
-            null,
-            UPLOAD_ERR_OK
-        );
-        $request = new Request([], ['operations' => '{"query": "graphqlQuery", "variables": {"file":null}, "operation": "graphqlOperationName"}', 'map' => '{"0":["variables.file"]}'], [], [], ['0' => $file]);
+        $requestParams = [];
+        if ($operations) {
+            $requestParams['operations'] = $operations;
+        }
+        if ($map) {
+            $requestParams['map'] = $map;
+        }
+        $request = new Request([], $requestParams, [], [], $files);
         $request->setMethod('POST');
         $request->headers->set('Content-Type', 'multipart/form-data');
 
         $schema = $this->prophesize(Schema::class);
         $schemaBuilderProphecy = $this->prophesize(SchemaBuilderInterface::class);
         $schemaBuilderProphecy->getSchema()->willReturn($schema->reveal());
+        $propertyAccessor = PropertyAccess::createPropertyAccessor();
 
         $executionResultProphecy = $this->prophesize(ExecutionResult::class);
         $executionResultProphecy->toArray(3)->willReturn(['GraphQL']);
         $executorProphecy = $this->prophesize(ExecutorInterface::class);
-        $executorProphecy->executeQuery(Argument::is($schema->reveal()), 'graphqlQuery', null, null, ['file' => $file], 'graphqlOperationName')->willReturn($executionResultProphecy->reveal());
+        $executorProphecy->executeQuery(Argument::is($schema->reveal()), 'graphqlQuery', null, null, $variables, 'graphqlOperationName')->willReturn($executionResultProphecy->reveal());
 
         $twigProphecy = $this->prophesize(TwigEnvironment::class);
         $routerProphecy = $this->prophesize(RouterInterface::class);
@@ -109,9 +115,65 @@ class EntrypointActionTest extends TestCase
         $graphiQlAction = new GraphiQlAction($twigProphecy->reveal(), $routerProphecy->reveal(), true);
         $graphQlPlaygroundAction = new GraphQlPlaygroundAction($twigProphecy->reveal(), $routerProphecy->reveal(), true);
 
-        $mockedEntrypoint = new EntrypointAction($schemaBuilderProphecy->reveal(), $executorProphecy->reveal(), $graphiQlAction, $graphQlPlaygroundAction, true, true, true, 'graphiql');
+        $mockedEntrypoint = new EntrypointAction($schemaBuilderProphecy->reveal(), $executorProphecy->reveal(), $graphiQlAction, $graphQlPlaygroundAction, $propertyAccessor, true, true, true, 'graphiql');
 
-        $this->assertEqualsWithoutDateHeader(new JsonResponse(['GraphQL']), $mockedEntrypoint($request));
+        $this->assertEquals($response->getContent(), $mockedEntrypoint($request)->getContent());
+    }
+
+    public function multipartFormProvider(): array
+    {
+        $file = new UploadedFile(
+          __DIR__.'/Fixtures/test.gif',
+          'original.gif',
+          null,
+          UPLOAD_ERR_OK
+        );
+
+        return [
+            'uploading a single file' => [
+                '{"query": "graphqlQuery", "variables": {"file":null}, "operation": "graphqlOperationName"}',
+                '{"file":["variables.file"]}',
+                ['file' => $file],
+                ['file' => $file],
+                new JsonResponse(['GraphQL']),
+            ],
+            'uploading multiple files' => [
+                '{"query": "graphqlQuery", "variables": {"files":[null,null,null]}, "operation": "graphqlOperationName"}',
+                '{"0":["variables.files.0"],"1":["variables.files.1"],"2":["variables.files.2"]}',
+                [
+                    '0' => $file,
+                    '1' => $file,
+                    '2' => $file,
+                ],
+                ['files' => [
+                    $file,
+                    $file,
+                    $file,
+                ]],
+                new JsonResponse(['GraphQL']),
+            ],
+            'uploading without providing operations' => [
+                null,
+                '{"file":["variables.file"]}',
+                ['file' => $file],
+                ['file' => $file],
+                new Response('{"errors":[{"message":"GraphQL query is not valid","extensions":{"category":"graphql"}}]}'),
+            ],
+            'uploading without providing map' => [
+                '{"query": "graphqlQuery", "variables": {"file":null}, "operation": "graphqlOperationName"}',
+                null,
+                ['file' => $file],
+                ['file' => null],
+                new JsonResponse(['GraphQL']),
+            ],
+            'upload with wrong map' => [
+                '{"query": "graphqlQuery", "variables": {"file":null}, "operation": "graphqlOperationName"}',
+                '{"file":["file"]}',
+                ['file' => $file],
+                ['file' => null],
+                new JsonResponse(['GraphQL']),
+            ],
+        ];
     }
 
     public function testBadContentTypePostAction()
@@ -150,6 +212,7 @@ class EntrypointActionTest extends TestCase
         $schema = $this->prophesize(Schema::class);
         $schemaBuilderProphecy = $this->prophesize(SchemaBuilderInterface::class);
         $schemaBuilderProphecy->getSchema()->willReturn($schema->reveal());
+        $propertyAccessor = PropertyAccess::createPropertyAccessor();
 
         $executionResultProphecy = $this->prophesize(ExecutionResult::class);
         $executionResultProphecy->toArray(3)->willReturn(['GraphQL']);
@@ -162,6 +225,6 @@ class EntrypointActionTest extends TestCase
         $graphiQlAction = new GraphiQlAction($twigProphecy->reveal(), $routerProphecy->reveal(), true);
         $graphQlPlaygroundAction = new GraphQlPlaygroundAction($twigProphecy->reveal(), $routerProphecy->reveal(), true);
 
-        return new EntrypointAction($schemaBuilderProphecy->reveal(), $executorProphecy->reveal(), $graphiQlAction, $graphQlPlaygroundAction, true, true, true, 'graphiql');
+        return new EntrypointAction($schemaBuilderProphecy->reveal(), $executorProphecy->reveal(), $graphiQlAction, $graphQlPlaygroundAction, $propertyAccessor, true, true, true, 'graphiql');
     }
 }